added sweetalert and functionality that sends email after deleting account

+++ database/migrations/2020_09_11_202239_cart_controller.php
@@ -14,7 +14,7 @@ class CartController extends Migration
     public function up()
     {
         Schema::create('cart_products', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigInteger('id');
             $table->integer('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->mediumText('products');
